+++ spec/models/carto/ldap/configuration_spec.rb
@@ -7,34 +7,43 @@ describe Carto::Ldap::Configuration do
   include_context 'organization with users helper'
 
 
-  it 'tests authentication' do
+  it 'tests basic authentication' do
+    domain_bases = [ "dc=cartodb" ]
 
-    auth_username = 'carto_admin'
+    ldap_admin_username = 'user'
+    ldap_admin_cn = "cn=#{ldap_admin_username},#{domain_bases[0]}"
+    ldap_admin_password =  '666'
+
+    auth_username = 'admin'
+    auth_cn = "cn=#{auth_username},#{domain_bases[0]}"
     auth_password = '123456'
-
-    user_domain_base = "cn=#{auth_username}"
-
-    domain_bases = [ user_domain_base, "dc=cartodb" ]
 
     ldap_configuration = Carto::Ldap::Configuration.create({
         organization_id: @organization.id,
         host: "0.0.0.0",
         port: 389,
         domain_bases: domain_bases,
-        connection_user: domain_bases,
-        connection_password: auth_password,
+        connection_user: ldap_admin_cn,
+        connection_password: ldap_admin_password,
         email_field: '.',
         user_object_class: '.',
         group_object_class: '.',
         user_id_field: 'cn'
       })
 
-    FakeNetLdap.register_user(:username => domain_bases.join(','), :password => auth_password)
+    FakeNetLdap.register_user(:username => ldap_admin_cn, :password => ldap_admin_password)
+
+    # This uses ldap_admin credentials
+    ldap_configuration.test_connection.should eq true
+
+    FakeNetLdap.register_user(:username => auth_cn, :password => auth_password)
 
     filter = Net::LDAP::Filter.eq('cn', auth_username)
-    FakeNetLdap.register_query(filter, { 
-          ldap_configuration.user_id_field => [auth_username]
-        })
+    FakeNetLdap.register_query(filter, 
+      # Data to return as an LDAP result, that will be loaded into Carto::Ldap::Entry
+      { 
+        ldap_configuration.user_id_field => [auth_username]
+      })
 
     result = ldap_configuration.authenticate(auth_username, auth_password)